no warning for type convert

+++ libcutils/include/cutils/bitops.h
@@ -45,15 +45,15 @@
 
 __BEGIN_DECLS
 
-static inline int popcount(unsigned int x) {
+static inline unsigned int popcount(unsigned int x) {
     return __builtin_popcount(x);
 }
 
-static inline int popcountl(unsigned long x) {
+static inline unsigned long popcountl(unsigned long x) {
     return __builtin_popcount(x);
 }
 
-static inline int popcountll(unsigned long long x) {
+static inline unsigned long long popcountll(unsigned long long x) {
     return __builtin_popcountll(x);
 }
